+++ modules/apps/portal-workflow/portal-workflow-kaleo-uad/src/main/java/com/liferay/portal/workflow/kaleo/uad/anonymizer/BaseKaleoInstanceTokenUADAnonymizer.java
@@ -59,11 +59,13 @@ public abstract class BaseKaleoInstanceTokenUADAnonymizer
 	}
 
 	@Override
-	public void delete(KaleoInstanceToken kaleoInstanceToken)
+	public void delete(KaleoInstanceToken kaleoInstanceToken, long userId)
 		throws PortalException {
 
-		kaleoInstanceTokenLocalService.deleteKaleoInstanceToken(
-			kaleoInstanceToken);
+		if (kaleoInstanceToken.getUserId() == userId) {
+			kaleoInstanceTokenLocalService.deleteKaleoInstanceToken(
+				kaleoInstanceToken);
+		}
 	}
 
 	@Override
