+++ modules/apps/portal/portal-uad/src/main/java/com/liferay/portal/uad/anonymizer/BaseRepositoryUADAnonymizer.java
@@ -57,8 +57,12 @@ public abstract class BaseRepositoryUADAnonymizer
 	}
 
 	@Override
-	public void delete(Repository repository) throws PortalException {
-		repositoryLocalService.deleteRepository(repository);
+	public void delete(Repository repository, long userId)
+		throws PortalException {
+
+		if (repository.getUserId() == userId) {
+			repositoryLocalService.deleteRepository(repository);
+		}
 	}
 
 	@Override
